services/memo_processor/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
anthropic==0.39.0
pymupdf4llm==0.0.17
python-docx==1.1.2
psycopg2-binary==2.9.10
boto3==1.35.0
httpx==0.28.0
python-dotenv==1.0.1
